Refactor App path
