+++ app/javascript/react/src/App.jsx
@@ -4,7 +4,7 @@ import {
     Routes,
     Route,
 } from 'react-router-dom';
-import Greeting from './Greeting';
+import Greeting from './components/Greeting';
 
 function App() {
     return (
